requirements.txt
absl-py==2.1.0
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bayesian-optimization==2.0.3
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.8.2
branca==0.7.2
bs4==0.0.1
cachetools==5.5.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-sdk==0.44.1
debugpy==1.8.2
DecisionTreeClassifier==0.0.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
docker==7.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.2.10
folium==0.17.0
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.4
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.0
JPype1==1.5.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.9.0
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==18.1.1
lxml==4.9.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mlflow==2.20.2
mlflow-skinny==2.20.2
mpmath==1.3.0
multitasking==0.0.11
mysql-connector-python==8.3.0
namex==0.0.8
narwhals==1.21.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==24.0
panda==0.3.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.1
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.24.1
prettytable==3.11.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.29.2
psutil==6.0.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycoingecko==3.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.4.2
scikit-surprise==1.1.4
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.34
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.1
tabula-py==2.9.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.6.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.13
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wget==3.2
widgetsnbextension==4.0.11
wrapt==1.17.2
xgboost==2.1.4
xyzservices==2024.6.0
yfinance==0.2.40
zipp==3.18.1